# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Intelligent AI Assistant API
  description: |
    Next-generation **Agentic AI Life Assistant** API designed to act as the user's conscious digital twin, 
    continuously learning, guiding, and optimizing real-world actions in real time.
    
    ## Features
    - **Voice Processing**: Speech-to-text and text-to-speech with Whisper and Coqui/ElevenLabs
    - **Calendar Integration**: Google Calendar synchronization with intelligent scheduling
    - **WhatsApp Messaging**: Business Cloud API integration for confirmations and updates
    - **Federated Learning**: Privacy-preserving machine learning with FedAvg and Differential Privacy
    - **Digital Avatar**: 2D/3D animated user representation during interactions
    - **Privacy Compliance**: GDPR, DIFC, and UAE Data Protection Law compliance
    
    ## Authentication
    All endpoints (except health checks and webhook verification) require JWT authentication.
    Include the Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Rate Limiting
    API requests are rate-limited to prevent abuse:
    - **Authentication endpoints**: 5 requests per minute
    - **Voice processing**: 20 requests per minute
    - **General API**: 100 requests per minute
    
    ## Error Handling
    The API uses standard HTTP status codes and returns structured error responses:
    ```json
    {
      "error": "Error description",
...
      "request_id": "uuid",
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      description: Create a new user account with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  minLength: 8
                  example: securepassword123
                timezone:
                  type: string
                  example: UTC
                language_preference:
                  type: string
                  enum: [en-US, fa-IR, ar-UA]
                  example: en-US
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid input or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user and return JWT tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  example: securepassword123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: Get new access token using refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
                  example: eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9...
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current user
      description: Get authenticated user information
      responses:
        '200':
          description: User information retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Voice Processing Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication.
        
        **Token Format:**
        ```
        Authorization: Bearer <jwt_token>
        ```
        
        **Token Lifecycle:**
        - Access tokens expire after 1 hour
        - Refresh tokens expire after 30 days
        - Automatic token rotation supported
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "Invalid email or password"
        correlation_id:
          type: string
          format: uuid
          description: Request correlation ID for debugging
          example: "550e8400-e29b-41d4-a716-446655440000"
        request_id:
          type: string
          format: uuid
          description: Unique request identifier
          example: "6ba7b810-9dad-11d1-80b4-00c04fd430c8"
        timestamp:
          type: number
          format: float
          description: Unix timestamp of error
          example: 1640995200.0
        details:
          type: object
          description: Additional error details
          additionalProperties: true
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
          example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9..."
        refresh_token:
          type: string
          description: JWT refresh token
          example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9..."
        token_type:
          type: string
          example: "bearer"
        expires_in:
          type: integer
          description: Access token expiration in seconds
          example: 3600
        user:
          $ref: '#/components/schemas/UserResponse'
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: "user@example.com"
        avatar_url:
          type: string
          format: uri
          nullable: true
          example: "https://api.ai-assistant.com/avatars/user_123.jpg"
        timezone:
          type: string
          example: "UTC"
        language_preference:
          type: string
          enum: [en-US, fa-IR, ar-UA]
          example: "en-US"
        created_at:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
...